+++ src/shared/components/ReactMapGl/map-style.js
