+++ src/shared/components/ReactMapGl/map-style.json
